Fix: 간혈적으로 헤더인풋값 누락버그 해결 및 뷰포트 사이즈트래커함수 커스텀훅으로 변경
