+++ components/layout/Header/HeaderInputBar.tsx
@@ -13,12 +13,11 @@ export default function HeaderInputBar() {
   const handleInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {
     setInput(e.target.value);
   };
-
-  const handleSubmit = (e: React.ChangeEvent<HTMLFormElement>) => {
+  const handleSubmit = async (e: React.ChangeEvent<HTMLFormElement>) => {
     e.preventDefault();
-    const newInput = input;
-    queryClient.setQueryData(["inputValue"], newInput);
+    queryClient.setQueryData(["inputValue"], input);
     setInput("");
+    await queryClient.refetchQueries(["inputValue"]);
     router.push("/searchpage");
   };
