+++ components/layout/Header/HeaderLogo.tsx
@@ -1,29 +1,13 @@
 "use client";
 
 import logo from "../../../public/images/logo.png";
+import useViewportTracker from "@/util/useViewportTracker";
 import Image from "next/image";
 import { useRouter } from "next/navigation";
-import { useEffect, useState } from "react";
 
 export default function HeaderLogo() {
   const router = useRouter();
-
-  // 뷰포트사이즈 감지 이벤트핸들러
-  const [viewportWidth, setViewportWidth] = useState(0);
-  useEffect(() => {
-    function handleViewportChange() {
-      setViewportWidth(window.innerWidth);
-    }
-    if (typeof window !== "undefined") {
-      setViewportWidth(window.innerWidth);
-      window.addEventListener("resize", handleViewportChange);
-    }
-    return () => {
-      if (typeof window !== "undefined") {
-        window.removeEventListener("resize", handleViewportChange);
-      }
-    };
-  }, []);
+  const viewportWidth = useViewportTracker();
 
   return (
     <div className="cursor-pointer ml-8">
